refactor: FSD Phase 6f — 単一画面専用の component・定数・utility を各 pages へ移す (FRESTYLE-172)
components/ サブディレクトリや constants / utils に残る単一画面専用の部品を、
その画面の Slice へ移す。

- dashboard/CompanyLearningPanel / DashboardStats / LearningCalendar
  → pages/home/ui
- settings/CompanyAiSettings → pages/settings/ui
- useCompanyAiSettings → pages/settings/model
- constants/glossary → pages/help/config
- utils/auth (getCognitoAuthUrl) → pages/login/lib
- SNSSignInButton → pages/login/ui

これで components/dashboard / components/settings / utils が空になり削除。
components 直下は MessageInput のみ(ask-ai 用・Phase 6h)。

pages 間 import が無いことを確認済み(境界 lint 緑)。既存テストの assert は
変更していない。テスト件数 1184 で不変。
