+++ frontend/src/pages/help/config/glossary.ts
