+++ frontend/src/pages/home/ui/CompanyLearningPanel.tsx
