+++ frontend/src/pages/home/ui/__tests__/CompanyLearningPanel.test.tsx
