+++ frontend/src/pages/home/ui/__tests__/DashboardStats.test.tsx
