+++ frontend/src/pages/login/ui/__tests__/SNSSignInButton.test.tsx
