+++ frontend/src/pages/settings/ui/__tests__/CompanyAiSettings.test.tsx
